+++ CHANGELOG.md
@@ -16,6 +16,7 @@ CHANGELOG
 
 ### Frontend
 - Show version including link to upstream in footer (#49).
+- Use Vue-Select for chosing the columns' fields (#50).
 
 #### Preview
 - Remove inout field text, does not work anyway.